moved copy function and changed headers where it was neccesary
