+++ modules/core/adjacent/include/nt2/core/functions/expr/cdiff.hpp
@@ -13,7 +13,7 @@
 #include <nt2/core/functions/cdiff.hpp>
 #include <nt2/core/functions/common/cdiff.hpp>
 
-#include <nt2/sdk/memory/copy.hpp>
+#include <nt2/memory/functions/copy.hpp>
 #include <nt2/core/container/dsl.hpp>
 #include <nt2/core/functions/of_size.hpp>
 #include <nt2/include/functions/firstnonsingleton.hpp>
