+++ modules/core/container/table/include/nt2/core/functions/table/construct.hpp
@@ -14,7 +14,7 @@
 #include <nt2/dsl/functions/terminal.hpp>
 #include <nt2/include/functions/numel.hpp>
 #include <nt2/sdk/memory/category.hpp>
-#include <nt2/sdk/memory/copy.hpp>
+#include <nt2/include/functions/copy.hpp>
 #include <boost/simd/memory/iterator_category.hpp>
 
 namespace nt2 { namespace ext
@@ -98,6 +98,31 @@ namespace nt2 { namespace ext
       static_cast<typename A0::nt2_expression&>(a0) = a1;
     }
   };
+
+// terminal = terminal for multi architectural
+ BOOST_DISPATCH_IMPLEMENT  ( construct_, tag::cpu_
+                            , (A0)(A1)
+                            , ((node_ < A0,nt2::tag::terminal_
+                                      , boost::mpl::long_<0>
+                                      , nt2::container::domain
+                                      >
+                              ))
+                               ((node_ < A1,nt2::tag::terminal_
+                                      , boost::mpl::long_<0>
+                                      , nt2::container::domain
+                                      >
+                              ))
+                            )
+  {
+    typedef void result_type;
+
+    BOOST_FORCEINLINE
+    result_type operator()(A0& a0, A1 const& a1) const
+    {
+      boost::proto::value(a0).assign(boost::proto::value(a1));
+    }
+  };
+
 
   /// INTERNAL ONLY
   /// Construct a terminal from a scalar:
