+++ modules/core/sdk/include/nt2/memory/functions/copy.hpp
@@ -6,8 +6,8 @@
 //                 See accompanying file LICENSE.txt or copy at
 //                     http://www.boost.org/LICENSE_1_0.txt
 //==============================================================================
-#ifndef NT2_SDK_MEMORY_COPY_HPP_INCLUDED
-#define NT2_SDK_MEMORY_COPY_HPP_INCLUDED
+#ifndef NT2_MEMORY_FUNCTIONS_COPY_HPP_INCLUDED
+#define NT2_MEMORY_FUNCTIONS_COPY_HPP_INCLUDED
 
 #include <boost/detail/workaround.hpp>
 #include <algorithm>
